Move network-related code to separate netcode-package
