+++ src/main/java/netcode/NetworkPackage.java
@@ -1,4 +1,4 @@
-package game;
+package netcode;
 
 import card.ICard;
